BE:fmt: organize files related to game logic
