+++ backend/src/game/game.py
@@ -1,0 +1,48 @@
+import src.models as d
+import src.schemas as s
+from src.game.deathmatch import Deathmatch
+
+
+class GameManager:
+    """
+    Manages active games, storing them in memory. Upon game finalization, the game is
+    removed from the manager and returned to be persisted in DB.
+    """
+
+    def __init__(self) -> None:
+        self.games: dict[int, Deathmatch] = {}
+
+    def get(self, game_id: int) -> Deathmatch:
+        return self.games[game_id]
+
+    def create(self, game_db: d.Game) -> Deathmatch:
+        if game_db.rules['type_'] == s.GameTypeEnum.DEATHMATCH:
+            rules = s.DeathmatchRules(**game_db.rules)
+            game = self.games[game_db.id_] = Deathmatch(
+                game_db.id_, game_db.players, rules
+            )
+            return game
+        else:
+            raise NotImplementedError('Unsupported game type')
+
+    def end(self, game_id: int) -> Deathmatch:
+        game = self.games[game_id]
+        game.status = d.GameStatusEnum.FINISHED
+        del self.games[game_id]
+        return game
+
+
+# TODO: Build and use abstract interface when you figure out the interface
+# class Game(Protocol):
+#     id_: int
+#     status: d.GameStatusEnum
+#     turns: list[s.Turn]
+
+#     def __init__(self, rules: s.Rules):
+#         ...
+
+#     def start_turn(self) -> s.TurnResults:
+#         ...
+
+#     def process_turn(self, word: str) -> s.TurnResults:
+#         ...
